# lib/db/src/schema/users.ts
import { pgTable, text, serial, timestamp, integer, boolean } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const usersTable = pgTable("users", {
  id: serial("id").primaryKey(),
  name: text("name").notNull(),
  email: text("email").notNull().unique(),
  passwordHash: text("password_hash").notNull(),
  role: text("role").notNull().default("user"),
  planId: integer("plan_id"),
  planExpiresAt: timestamp("plan_expires_at", { withTimezone: true }),
  /** Marketing / admin-granted premium trial — overrides tier until this instant */
  premiumTrialEndsAt: timestamp("premium_trial_ends_at", { withTimezone: true }),
  creditsUsed: integer("credits_used").notNull().default(0),
  creditsLimit: integer("credits_limit").notNull().default(5),
  dailyCreditsUsed: integer("daily_credits_used").notNull().default(0),
  dailyLimit: integer("daily_limit").notNull().default(5),
  dailyResetAt: timestamp("daily_reset_at", { withTimezone: true }),
  isSuspended: boolean("is_suspended").notNull().default(false),
  resetPasswordToken: text("reset_password_token"),
  resetPasswordExpires: timestamp("reset_password_expires", { withTimezone: true }),
  createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
  updatedAt: timestamp("updated_at", { withTimezone: true }).notNull().defaultNow().$onUpdate(() => new Date()),
});

export const insertUserSchema = createInsertSchema(usersTable).omit({ id: true, createdAt: true, updatedAt: true });
export type InsertUser = z.infer<typeof insertUserSchema>;
export type User = typeof usersTable.$inferSelect;